List all known URL sources for a given Content URI

List all known URL sources for a given Content URI

query_sources(
  id,
  registries = default_registries(),
  cols = c("source", "date"),
  ...
)

Arguments

id
a content identifier

registries
list of registries at which to register the URL

cols
names of columns to keep. Default are source and date. See details.

...
additional arguments

Value

a data frame with all registration events when a URL or a local path (including the local store) have contained the corresponding content.

Details

possible columns are (in order): identifier, source, date, size, status, md5, sha1, sha256, sha384, sha512

See also

Examples

# \donttest{

id <- paste0("hash://sha256/9412325831dab22aeebdd",
             "674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37")
query_sources(id)

#>                                                                                                                            source
#> 1               /home/cboettig/.local/share/contentid/data/94/12/9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37
#> 2 https://archive.softwareheritage.org/api/1/content/sha256:9412325831dab22aeebdd674b6eb53ba6b7bdd04bb99a4dbb21ddff646287e37/raw/
#> 3                                  https://knb.ecoinformatics.org/knb/d1/mn/v2/object/ess-dive-457358fdc81d3a5-20180726T203952542
#> 4                               https://knb.ecoinformatics.org/knb/d1/mn/v2/object/ess-dive-f8d52384ece0ba9-20180716T154049822990
#> 5                                          https://github.com/espm-157/climate-template/releases/download/data/vostok.icecore.co2
#> 6                                             https://cn.dataone.org/cn/v2/resolve/ess-dive-0462dff585f94f8-20180716T160600643874
#> 7                                                                https://cdiac.ess-dive.lbl.gov/ftp/trends/co2/vostok.icecore.co2
#> 8                                                                      https://zenodo.org/record/3678928/files/vostok.icecore.co2
#> 9                            https://data.ess-dive.lbl.gov/catalog/d1/mn/v2/object/ess-dive-642bb2e5bd48452-20180716T160320693400
#>                  date
#> 1 2020-07-29 21:11:18
#> 2 2020-08-18 03:33:12
#> 3 2020-08-17 23:51:15
#> 4 2020-08-10 00:09:02
#> 5 2020-08-04 18:25:50
#> 6 2020-08-01 00:01:40
#> 7 2020-07-31 21:05:59
#> 8 2020-07-10 23:15:00
#> 9 2020-07-10 23:14:18

# }